feat: limit registry nesting to 2 levels for registry sources (#7094)
## Description


Limits the levels of nesting to 2 for registry sources. Changes the
naming convention for the index files from unlimited amount of nesting
by `chunk_size` to only 2 levels of nesting with `chunk_size`.

So `foobar` was split as `fo/ob/ar/foobar` with `chunk_size` 2 but with
this PR it is split as `fo/ob/foobar`.