Remove `AbiEncode` impl for `raw_ptr`
This was added originally in https://github.com/FuelLabs/sway/pull/6844
but should not have been introduced in the first place.

I added this while in development for the trait coherence feature but
its not necessary and not something we wish to support in the public ABI
surface.

Fixes https://github.com/FuelLabs/sway/issues/7153.