Implement a command to dump impls for a given type
This adds a new `--dump-impls` CLI command that can be used to dump all
the recognized trait impls for a given type name.

This also extends the compiler diagnostic system to allow the emission
of info diagnostics, as previously we could only emit either warnings of
errors, which did not fit this.

Closes https://github.com/FuelLabs/sway/issues/7286.