Don't duplicate source pointers in spans
When doing multiple subsequent compilations, such as when running tests
or when using the LSP, and reusing engines, in particular the decl engine,
an condition can occur when files are loaded a second time and generate new
spans that point to a different string than existing spans for that file.

This creates a problem when joining spans that point to the same source
file but not the same string.

This PR adds a new functionality to the source engine, where it will
store pointers to source strings (that spans eventually use) and
deduplicate them before they can cause this issue.

A future PR may want to reuse this facility to prevent unnecessary file
loads altogether.

Fixes #7338