Improve method resolution by allowing placeholder unification under ConstraintSubset (#7411)
## Description

Closes https://github.com/FuelLabs/sway/issues/7398.

The issue here is that `UnificationCheck` was not considering
`Placeholder` under `ConstraintSubset`, the strategy that is used for
method resolution.

This was affecting the test `trait_generic_override` because
`MyStruct<Placeholder(T)>` was not unifying with `MyStruct<u64>`; What
was happening was that `check_struct` was calling
`check_multiple([Placehodler(T)], [u64])`, which is correct; but the
`match` trying to unify them under `ConstraintSubset` was falling into
the wildcard arm, and end up caomparing both types with `eq`.

To "fix" this "unification failure", we implemented the `if` under the
comment of "Consider items from supersets indicated by the annotation
return type". This was working for this example, but miscompiling other
cases.

This PR is trying a better approach which is improving the unification
and removing that "if".

After these changes, the compiler started to give a wrong advice as in
the example below:

```
  --> test/src/e2e_vm_tests/test_programs/should_fail/trait_constraint_generic_hint/src/main.sw:43:13
   |
41 |     }
42 | 
43 |     if item.method() {
   |             ^^^^^^ Multiple applicable items in scope. 
  Disambiguate the associated function for candidate #0
    <bool as A>::method
  Disambiguate the associated function for candidate #1
    <bool as B>::method
44 | 
45 |     }
   |
```

Currently there is no way to fix this ambiguity. So my suggestion is
that the compiler only complain about "Multiple applicable items in
scope" and we open another issue to improve this error message.
(https://github.com/FuelLabs/sway/issues/7398?issue=FuelLabs%7Csway%7C7414)

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [ ] If my change requires substantial documentation changes, I have
[requested support from the DevRel
team](https://github.com/FuelLabs/devrel-requests/issues/new/choose)
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [ ] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.