fix: also create deployment artifacts for chunked deployments (#7426)
## Description

The deployment artifact generation we have right now was not creating
the artifact for chunked deployments. This PR adds that as much as
possible as some fields are not expoed via sdk right now (the block
height and transaction id) for chunked deployments.

Once/if those are exposed we can extend this easily to add those fields
as well. But nevertheless this PR serializes majority of the information
we have about the chunked deployments to the `out/deployments` folder.

---------

Co-authored-by: Joshua Batty <joshpbatty@gmail.com>